fix: Added additional check in deleteLobby Method in GameLobbyEntityServiceImpl

+++ src/main/java/at/aau/serg/websocketdemoserver/service/impl/GameLobbyEntityServiceImpl.java
@@ -43,7 +43,12 @@ public class GameLobbyEntityServiceImpl implements GameLobbyEntityService {
 
     @Override
     public void deleteLobby(Long id) {
-        gameLobbyEntityRepository.deleteById(id);
+        if(gameLobbyEntityRepository.findById(id).isPresent()) {
+            gameLobbyEntityRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("gameLobby does not exist and can not be deleted");
+        }
+
     }
 
     @Override
